DataEngine/requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
python-dotenv>=1.0.0
aiohttp>=3.8.0